+++ apps/web/src/app/admin-console/organizations/settings/organization-settings-routing.module.ts
@@ -54,7 +54,7 @@ const routes: Routes = [
           {
             path: "export",
             loadChildren: () =>
-              import("../tools/vault-export/org-export.module").then(
+              import("../tools/vault-export/org-vault-export.module").then(
                 (m) => m.OrganizationVaultExportModule
               ),
           },
