+++ apps/web/src/app/admin-console/organizations/tools/vault-export/org-vault-export-routing.module.ts
@@ -5,7 +5,7 @@ import { Organization } from "@bitwarden/common/admin-console/models/domain/orga
 
 import { OrganizationPermissionsGuard } from "../../guards/org-permissions.guard";
 
-import { OrganizationVaultExportComponent } from "./org-export.component";
+import { OrganizationVaultExportComponent } from "./org-vault-export.component";
 
 const routes: Routes = [
   {
